Feat/back/image (#52)
* fix: create documento bugfix

* feat: Add endpoint for updating user photo

The `DocumentoController` has been updated to include a new endpoint `update_photo` for updating the user's photo. This endpoint accepts the `_id` of the user and the uploaded image file. The `DocumentoService` and `DocumentoRepository` have also been updated to handle the photo update logic.

* create bugfix

* feat: Add update_photo endpoint to AssuntoController and DocumentoController

The AssuntoController and DocumentoController have been updated to include a new endpoint called update_photo. This endpoint allows users to update their photo by providing the _id of the user and the uploaded image file. The AssuntoService and AssuntoRepository have also been updated to handle the photo update logic. This change improves the functionality of the application by allowing users to easily update their profile photos.

---------

Co-authored-by: Alexandre Beck <alexandrebeck2003@gmail.com>